[core] Deflake `test_output.py::test_dedup_logs` (#53148)
Test was flaky, even locally, because the timing in CI might cause the
tasks to run spaced out and therefore the logs not to be deduped.
Example flake:
https://buildkite.com/ray-project/postmerge/builds/10232#0196e94c-c1f5-494e-9ffd-232334fa8926/179-1408.

I've updated it to control the timing conditions more closely using a
`SignalActor`.

Before the change, it failed about 50% of the time on my laptop. After
the change, it passed 10/10 consecutive runs.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>