[core][compiled graphs] Fix execution schedules with collective operations (#53007)
Given an input DAG of SPMD training strategies such as DDP, after DAG
compile, the first actor will generate different execution schedules
than others. This is due to the current scheduling policy, when there
are multiple ready operation nodes such as `actor1.compute` (non-NCCL)
and `actor4.collective` (NCCL, for actor1-4, there's only one collective
operation node that's eventually ready), the policy does not know actor1
has both the non-NCCL `actor1.compute` and the NCCL `actor4.collective`.
This leads to actor1 scheduling the `actor1.compute` first, and actor1-4
scheduling the `collective` next.

We update the policy to push all the collective operations nodes into
candidates when the last of them is ready. In the previous example,
actor1 will have both `actor1.compute` and `actor1.collective` as
candidates. In a DAG of SPMD strategies, all the actors pop either the
`compute` or the `collective` together.

We also update the policy to simply prioritize the NCCL operation node
over the non-NCCL. This will lead to NCCL operations to be scheduled as
soon as possible. It is safe to do so under the current settings of CUDA
streams in the system, because each NCCL read/write/collective stream
only allows one outstanding NCCL kernel at a time.

We add a test `test_collective_dag.py::test_exec_schedules_ddp` to
verify the generated schedules are identical across workers for the DDP
stragegy. Other tests are updated to reflect the changes of prioritizing
the NCCL operation node over the non-NCCL.

## Related issue number

<!-- For example: "Closes #1234" -->

This PR is part of #48649 planning to be merged incrementally.

---------

Signed-off-by: Weixin Deng <weixin@cs.washington.edu>