[core][metri-agg/1] recommended level for last-value data (#52846)
This series of PR addresses
https://github.com/ray-project/ray/issues/47289. The context is that
there are metrics such as `ray_tasks` and `ray_actors` produce a high
volume of time series on prometheus. Our inspection indicates that this
is because the high cardinality of the `WorkerId` field in these metrics
in a high scale cluster. See
https://docs.google.com/document/d/1AZVZQGGroSbV1w4KG4Vncc0L_tVdpfkM3ueGo_fD9-M/edit?tab=t.0
for the proposed solution.

This PR:
- Add RAY_core_metric_cardinality_level which can be set to LEGACY,
RECOMMENDED or LOW as proposed in the google doc solution
- Implemented the RECOMMENDED cardinality for all metrics of type
LastValueAggregationData
- Add test cases

Subsequent PRs will implement additional features to fill the other
missing pieces. The PR is safe to merged in pieces because it does not
change the default behaviors (and the additional features are not
communicated publicly yet so will only be controlled by us).

Test:
- CI (test cases)
- e2e test in production as well

Signed-off-by: can <can@anyscale.com>