[core][metri-agg/2] recommended for all (mostly) data type (#52947)
This series of PR addresses
https://github.com/ray-project/ray/issues/47289. The context is that
there are metrics such as ray_tasks and ray_actors produce a high volume
of time series on prometheus. Our inspection indicates that this is
because the high cardinality of the WorkerId field in these metrics in a
high scale cluster. See
https://docs.google.com/document/d/1AZVZQGGroSbV1w4KG4Vncc0L_tVdpfkM3ueGo_fD9-M/edit?tab=t.0
for the proposed solution.

This PR:
- Small step to complete the RECOMMENDED aggregation level, by further
support all metric types except for distributed data. Note that
aggregation of distributed data at the node level from worker level are
not well defined (or very complicated to be well defined), and the
existing set of distributed data are already low cardinality, so I don't
spend too much time to support this type of data here. Can be adjusted.

Test:
- CI

Signed-off-by: can <can@anyscale.com>