[core] Improve actor GC error message to suggest storing the actor handle before use (#53172)
This PR improves the runtime error message raised when an actor handle
is garbage-collected during a remote method invocation, making it
clearer what went wrong and how to work around it.

Previously, invoking an actor method in a single expression like:
```
ray.get(MyActor.remote().do_work.remote())
```
would raise a generic error
```
RuntimeError: Lost reference to actor
```
because Ray only held a weak reference to the actor stub. Users were
left guessing how to avoid this error.

Note: this PR does not fix the actual issue. See
https://github.com/ray-project/ray/issues/6265 for more details. A way
to fix the issue is by keeping a strong reference to the actor but for
the brief window when we build and execute the invocation closure. Until
a more permanent fix is implemented, let's improve the error message
with the workaround to unblock users.

Closes https://github.com/ray-project/ray/issues/6265

Signed-off-by: sagar <sagar@anyscale.com>