[dashboard][train] Add dynolog for on-demand GPU profiling for Torch training (#53191)
Add on-demand GPU profiling endpoint to the dashboard reporter API at
`/worker/gpu_profile`.

Here's the list of parameters that are available:
```
/worker/gpu_profile?node_ip=xxx.x.x.x&pid=xxxxx&num_iterations=x
```

## Design

* Functionality depends on dynolog binaries (`dynolog`, `dyno`) being
installed on the image already.
* Launch a singleton dynolog daemon monitoring process on every node
during dashboard agent startup, listening at port `65406`.
* The `/worker/gpu_profile` request gets propagated to the ReporterAgent
on the correct node, which makes a call to `dyno gputrace
--pids=<train_worker_pid> ...` and then waits for the trace file to be
dumped.
* The request then redirects to the streaming log download API to
download the trace on the client (browser).

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>