[core] Deflake `test_memory_pressure.py` (#53237)
The test is failing periodically in CI due to memory usage being above
the memory monitor threshold prior to running an actor:
https://buildkite.com/ray-project/postmerge/builds/10292#0196f473-d8f0-41fa-8041-addf42660c57/179-477

I'm attempting to deflake it by converting it to follow the same pattern
as its sister test and combining them using a fixture. This only
requires that we're below the memory monitor threshold to start, not
that we are below `0.3 * threshold`.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>