Reduce flakiness of UI tests (#53215)
Attempts to fix two sources of flakiness:
1. `TypeError: Cannot read properties of null (reading
'visibilityState')` error message. We do "visibility" checks not because
we want to verify the document is visible, but because we want to verify
the element is visible. JSDom seems to sometimes not support the
visibilityState field but its not something we really need to verify, so
we mock it out.
2. Timeout for `ActorTable.component.test.tsx`. This is a long running
test because it involves a lot of UI element interaction. We increase
the timeout


Signed-off-by: Alan Guo <aguo@anyscale.com>