[core] Deflake `test_reference_counting.py` (#53332)
Since converting to a shared fixture, the test has been
[flaking](https://buildkite.com/ray-project/postmerge/builds/10401#01970e33-da98-4806-8975-111a279ed170/177-1248)
in CI.

I tracked this down to a cyclic reference in the failure condition of
`test_basic_serialized_reference` that was causing a reference to be
"leaked" until the garbage collector runs. I tried reproducing it
outside of the `pytest` setup and was not able to, so might be a quirk
related to `pytest`.

Deflaking by:

- Adding a `gc.collect()` inside the ref count checking logic.
- Adding a fixture to assert that all refs are cleared between tests to
help avoid & track down future issues.

Closes https://github.com/ray-project/ray/issues/43983

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>