[core] GRPCServer takes ownership of GRPCService (#53300)
The GRPCServer should take ownership of GrpcService. This way the class
that has the grpc server doesn't have to hold a grpc service for all of
its services to keep them alive for the duration of the service. This
decreases the amount of member variables in the class, simplifies
lifetimes, and makes dependency injection easier in the future since you
only have to inject a mock server, not mock services.

Signed-off-by: dayshah <dhyey2019@gmail.com>