[ci] Run flaky tests once per build (down from twice) (#53364)
We currently have many flaky tests, especially on Windows. This causes
the builds to be extremely long and sometimes time out.

This change reduces the number of baseline runs per CI job from 1 to 2.
This should ~halve the run time for this build at the cost of it taking
two CI jobs instead of one to promote a test out of flaky status.

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>