[core] Fix reference counter crashes during worker graceful shutdown (#53002)
## Problem description

See https://github.com/ray-project/ray/issues/53012 for minimal
reproduction of the issue.

Currently the shutdown sequence looks approximately like:

1. Blindly decrement all local reference counts.
2. Drain ongoing _submitted_ tasks.
3. Drain the local task executor. This drains ongoing _executed_ tasks
if the worker is part of a concurrent actor. For single-threaded actors
and task workers, it's basically a no-op.
4. Drain the reference counter.

Step (1) essentially invalidates the state of the reference counter,
meaning that subsequently executing any user code that interacts with
the reference counter will cause undefined behavior and/or crashes. This
is the problem observed in the two linked issues. In both cases, the
worker is executing concurrent actor tasks which are not drained until
step (3).

## Original motivation for `ReleaseAllLocalReferences`

The original motivation for this code is documented in:
https://github.com/ray-project/ray/issues/19639. However, I haven't been
able to fully understand the rationale or the fix. I also haven't been
able to reproduce the issue described, even when substituting the driver
for an owner actor to avoid job-level worker cleanup.

What happens in the described setup inside the worker is (confirmed
empirically):

1. Before the worker starts to execute the task, we [increase the local
reference
count](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/core_worker.cc#L3762)
for the argument as part of
[GetAndPinArgsForExecutor](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/core_worker.cc#L3739).
2. The worker executes the task, blocks in `time.sleep`, then raises
`SystemExit` when it receives a `SIGTERM`.
3. After `SystemExit` is raised, the rest of `ExecuteTask` is executed,
which [decrements the local reference
count](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/reference_count.cc#L968)
as part of
[PopAndClearLocalBorrowers](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/core_worker.cc#L3403).
4. The worker calls
[CoreWorker::Exit()](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/core_worker.cc#L3455).

During the shutdown sequence in (4), the reference has already been
cleared, so the call to `ReleaseAllLocalReferences` is a no-op.

However, there is another edge case that `ReleaseAllLocalReferences`
does currently handle, which is when a normal task (not an actor task)
stores an `ObjectRef` in a global variable before exiting. We currently
[assume this never
happens](https://github.com/ray-project/ray/blob/871cc89ceda42caaa4819112343ff7b0e005e0e4/src/ray/core_worker/core_worker.cc#L1229).
In this case, the call to `ReleaseAllLocalReferences` will decrement the
globally-held reference(s) and allow the worker to exit.

## Solution and alternatives

Originally I planned to fully delete the call to
`ReleaseAllLocalReferences`, but it would cause a behavior change in the
edge case described above pertaining to `ObjectRef`s stored in global
variables.

Therefore, I'm opting for a safe option for now: move the call to
`ReleaseAllLocalReferences` _after_ we drain the task executor, which
should be a no-op change for single-threaded actors and normal task
workers, but solve the issue of tasks executing after the reference
counter has been invalidated in the case of concurrent actors.

## Follow-up items

- We should solve the global `ObjectRef` issue for tasks separately,
then remove the `ReleaseAllLocalReferences` logic. (TODO: file ticket
before merging -- 1 ref and multiple refs)

Closes https://github.com/ray-project/ray/issues/51408
Closes https://github.com/ray-project/ray/issues/53012

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>