[core][telemetry/02] opentelemetry recorder grpc service (#53158)
This is a series of PR to migrate metric collection from opencencus to
openlemetry. For context on the existing components, see
https://github.com/ray-project/ray/pull/53098.

------------

This PR add the open telemetry grpc service to the DashboardAgent
component. The grpc service is behind a feature flag so this PR by
default is a no-op, but it allows me to make progress in a piece meal
fashion.

Test:
- CI
- e2e testing on top of this PR

---------

Signed-off-by: can <can@anyscale.com>