[core] Early return and cleanup around WarnResourceDeadlock (#53308)
When looking for a resource deadlock, the first thing done is to check
if there's any workers available. Right now that involves getting a
vector of all the workers and then checking.
Created a new worker pool IsAvailableForScheduling that will just return
out if there's one available.

AnyPendingTasksForResourceAcquisition (also called to check for resource
deadlock) was a kind of messy with a return value and four out
parameters. Simplified down to a return value and two out parameters.

The moves in FillResourceUsage weren't actually moving anything because
it was using the const protobuf accessors. Fixed that.

Signed-off-by: dayshah <dhyey2019@gmail.com>