[core][compiled graphs] Unify scheduling for NCCL operation nodes (#53111)
This PR unifies the scheduling implementation for the NCCL P2P and
collective operation nodes. The logic remains the same: (1) P2P case:
When a NCCL send node is selected, its downstream NCCL recv nodes are
also selected; (2) Collective case: When a NCCL collective node is
selected, its corresponding NCCL collective nodes are also selected.
Previously, the NCCL P2P case was implemented by selecting the recv
nodes if a send node is detected, and the NCCL collective case was
implemented by maintaining a set of pending collective nodes.

We unify the implementation for both cases. Concretely, they both
maintain a set of (pending) synchronous nodes named `sync_idxs` and
`pending_sync_idxs`. The synchronous nodes denote the P2P send/recv
nodes or the collective nodes. The NCCL P2P/collective operation is
ready when `sync_idxs == pending_sync_idxs`.

Test cases are updated to reflect the use of synchronous nodes for both
NCCL P2P and collective nodes.

This PR is a follow-up of #53007. They are parts of #48649 planning to
be merged incrementally.
---------

Signed-off-by: Weixin Deng <weixin@cs.washington.edu>