[core][telemetry/03] add OpenTelemetryMetricRecorder c++ client (#53159)
This is a series of PR to migrate metric collection from opencencus to
openlemetry. For context on the existing components, see
https://github.com/ray-project/ray/pull/53098.

------------

This PR add a `OpenTelemetryMetricRecorder` class that has the
architecture very similar to its python counter part in
https://github.com/ray-project/ray/pull/53098.
- It is also a singleton class and will hold the snap shot of all
metrics throughout the process lifecycle
- The main difference is, this uses a meter that push to a grpc endpoint
vs. a prometheus-pull meter like the python counter part

Test:
- CI
- e2e testing on top of this PR

---------

Signed-off-by: can <can@anyscale.com>
Signed-off-by: Cuong Nguyen <128072568+can-anyscale@users.noreply.github.com>
Co-authored-by: Dhyey Shah <dhyey2019@gmail.com>