[data/preprocessors] Add flatten functionality to Concatenator (#53378)
Addresses #51757 

Adds a `flatten` boolean flag to the `Concatenator` preprocessor, which
will flatten interior vectors when concatenating, for example when you
use a OneHotEncoder.

I added tests for basic flattening functionality, verifying flatten=True
vs False, as well as a special case with dtype coercion we found when
testing with our own users internally.

I used the precommit steps for linting et al - I ultimately had to
ignore pydoclint due to it complaining about typehints in the doc not
matching - but when I added typehints, it complained that typehints in
the docs are not supported. It also generated a huge change to the doc
baseline file that seemed irrelevant to the PR - let me know if I should
do something about this, but it seems this pydoclint would fail even
before my changes. Happy to fixup with some guidance in a follow up
commit.

## Why are these changes needed?

Details in #51757 - Concatenator doesn't work as expected once we
changed encoders to output vectors in a single column. This adds a
feature to support the previous expectations of the output of
concatenator providing a flat vectorizable output.

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Ryan Clough <rclough@spotify.com>
Signed-off-by: Richard Liaw <rliaw@berkeley.edu>
Co-authored-by: Richard Liaw <rliaw@berkeley.edu>