[core][gpu-objects] GPU Objects POC (#52938)
![image](https://github.com/user-attachments/assets/68103cde-1f73-4354-8ba1-ce85760696fe)

This PR implements most parts of the diagram above.
* Currently, this PR supports only Tensors, not TensorDicts.
* This PR supports data that mixes both CPU and GPU.
* This PR implements a very simple GC mechanism: all data in the
in-actor object store is garbage-collected once it has been consumed.

### Details
* Step 1: Users annotate the sender’s actor method with
`@ray.method(tensor_transport=...)`. The valid, case-insensitive values
for `tensor_transport` are `nccl`, `gloo`, and `object_store` (default).
  ```python
  @ray.remote
  class GPUTestActor:
      @ray.method(tensor_transport="GLOO")
      def echo(self, data):
          return data
      ...
  ```
* Step 2: Users create a communication group, such as an NCCL group, for
actors that need to communicate with each other. In addition, each actor
needs to register the custom serializer.
  * This step can be optimized for better UX in the future.

* Step 3: Pass the tensor_transport information through the stack—Python
→ Cython → C++ → Cython —when submitting a task to the sender actor.
  * Python (actor.py) → Cython → C++ → Cython
* If `tensor_transport` is not `OBJECT_STORE`,
`serialize_and_store_gpu_objects` will be called to extract tensors from
the task output and store them in the `GPUObjectManager`.

* Step 4: When the driver process resolves the dependencies of the
receiver actor’s task argument, if that argument is an `ObjectRef`
pointing to an object created by an actor method annotated with
`@ray.method(tensor_transport="...")` (NCCL or GLOO), it submits a
`__ray_send__` task to the sender actor to initiate the send operation
(e.g., NCCL send) and a `__ray_recv__` task to the receiver actor to
initiate the receive operation (e.g., NCCL recv).

* Step 5: Pass the object ID through the stack—C++ (driver) → C++
(receiver actor) → Cython → Python (`def deserialize`).
* In step 4, the data’s tensors are transferred via out-of-band NCCL or
GLOO communication, then stored in the receiver actor’s in-actor object
store using the object ID as the key.
* In `def deserialize`, use the object ID to retrieve tensors from the
in-actor object store, add them to the serialization context, and then
deserialize to obtain the argument.

---------

Signed-off-by: Stephanie wang <smwang@cs.washington.edu>
Signed-off-by: Kai-Hsun Chen <kaihsun@anyscale.com>
Signed-off-by: Kai-Hsun Chen <kaihsun@apache.org>
Co-authored-by: Stephanie wang <smwang@cs.washington.edu>
Co-authored-by: Stephanie Wang <swang@cs.berkeley.edu>