[core] Revert KillActor rpc failure error log (#53412)
Reverting the error log added here
https://github.com/ray-project/ray/pull/53321/files.

Things were worse than I imagined. The success behavior for KillActor is
the rpc failing. If the actor is not running on the worker, we'll send
invalid status. If the actor kills

I'll rework this rpc in a follow up. There's a couple different ideas
1. gcs checks if the actor is dead -> gcs checks if the node is dead ->
asks the raylet if the worker is dead -> retries if all are false.
2. gcs tells raylet to kill the actor worker -> since it's
raylet->worker is ipc, the reply dying before the reply makes it in
force exits is less likely. Normal gcs -> raylet retries with backoff
that check if the node is alive before retrying.

Prefer option 2 right now.


https://github.com/ray-project/ray/blob/7e4d2ecb40cf41fe39d2a5a44e9fa934735e535d/src/ray/core_worker/core_worker.cc#L4522-L4551


Signed-off-by: dayshah <dhyey2019@gmail.com>