[Core] Add bundle_label_selector scheduling logic (#52988)
This PR adds the `bundle_label_selector` option to Ray PlacementGroups,
applying the label selectors on a per-bundle level to define label
constraints for scheduling bundles on a node.

This PR passes the `bundle_label_selector` from the `PlacementGroup`
options to the Raylet where the `PlacementGroupSpec` is built. Inside
`PlacementGroupSpec`, each label selector is added to its corresponding
`Bundle` message, and then added to the `ResourceRequest` in
`ComputeResources` when the `BundleSpec` is constructed. This enables
the cluster resource scheduler to check for required labels when calling
`IsFeasible` and `IsAvailable` on a Node.

https://github.com/ray-project/ray/issues/51564

---------

Signed-off-by: Ryan O'Leary <ryanaoleary@google.com>
Co-authored-by: Dhyey Shah <dhyey2019@gmail.com>
Co-authored-by: Mengjin Yan <mengjinyan3@gmail.com>