[tune] relax test_experiment_restore timeout (#53387)
This test is flakily timing out, in particular for Python 3.12 tests.
This is most likely due to minor regressions that have built up over
time across the stack, and perhaps test concurrency?

Increasing this buffer for now. In the future we may want to improve
this test to be more deterministic and restrictive in testing.

---------

Signed-off-by: Matthew Deng <matt@anyscale.com>