[core] Handle node death in object manager (#53397)
The main issue here is that if a node dies after a pull request is made.
The node that got the pull request will try to push the object to
completion. Pushing large objects can be slow, especially if they're on
disk. And pushing objects to dead nodes means that the rpc threads will
be busy trying to push to dead nodes instead of doing the work of
pushing to alive nodes.

1. When a node dies, it's death will get propagated to the object
manager and the push manager.
The object manager will remove any client it has for that node.
The push manager will remove all the push states for that node id.

2. Giving the object manager its own gcs_cilent so it doesn't need to
use the object directory as a proxy to the gcs client. This doesn't make
any sense and was probably a relic of old code structure.

3. The push manager push state map is now a map of node id -> map of obj
id -> an iterator to the list. And the list is now a list of PushState.
This way when a node is removed, we can directly erase from the list and
the map. There's some reworking in PushManager to make this happen, but
in general the logic should still be the same.

4. Before, the push manager erased PushState when the reply to the last
push request is received. Now, the push manager erases the PushState
when the last push request is sent out. This changes the
`push_manager_in_flight_pushes` metric which was the # of PushState's to
`push_manager_num_pushes_remaining`, still the # of PushState's.

---------

Signed-off-by: Dhyey Shah <dhyey2019@gmail.com>
Co-authored-by: Edward Oakes <ed.nmi.oakes@gmail.com>