[core][telemetry/04] register and record metrics on worker side (#53209)
This is a series of PR to migrate metric collection from opencencus to
openlemetry. For context on the existing components, see
https://github.com/ray-project/ray/pull/53098.

--------------

This PR add logics to register and set value for a gauge metric. It also
registering the `OpenTelemetryMetricRecorder` with the grpc endpoint.
This integration allows me to test things end-to-end but is a safe no-op
because it's behind a feature flag.

Test:
- CI
- build compiled

---------

Signed-off-by: can <can@anyscale.com>
Signed-off-by: Cuong Nguyen <128072568+can-anyscale@users.noreply.github.com>
Co-authored-by: Dhyey Shah <dhyey2019@gmail.com>