[core] Speed up and deflake `test_output.py` (#53584)
Flakiness observed in postmerge:
https://buildkite.com/ray-project/postmerge/builds/10600#0197346f-6470-4382-b617-be7b5ac6e4e1/177-1531

Various cleanups bundled in this PR:

- Combine two redundant tests for autoscaler log messages.
- Add `RAY_debug_dump_period_milliseconds` to relevant test to speed up
time it takes for infeasible message to be logged (previously it was
25s, now 5s).
- Incrementally check for log output in a few tests rather than sleeping
blindly for a configured number of seconds.
- Remove parametrization for autoscaler v2 event test to share the same
Ray instance.
- Remove parametrization for stdout/stderr in task & actor output tests.
- Use explicit signaling with a temp file instead of `time.sleep`
waiting for a driver to start in `test_output_on_driver_shutdown`.
- Avoid running `test_empty_line_thread_safety_bug` 3x per build and
refactor it to avoid sleeping.
- Remove `test_node_name_in_raylet_death`: all this is doing is
asserting on the contents of a log line deep in the GCS... 😕

Local e2e test times:

- Before: `37 passed in 366.20s (0:06:06)`
- After: `22 passed in 133.00s (0:02:13)`

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>