Fix python executable path for ray nsight configuration (#53598)
## Why are these changes needed?

Not all systems are guaranteed to have "python" be the way that the
python interpreter is invoked on the command line. The proper way to
find python is to use `sys.executable`. This hardcoded "python" was
breaking workflows within my Ubuntu 22.04 training environment.

## Related issue number

Closes #53597 

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [x] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [x] This PR is not tested :(

---------

Signed-off-by: Shishir Bhat <shishirvbhat@gmail.com>