[core] Move some utils to `ray._common.utils` to replace `ray._private.utils` usage in libraries (#53287)
## Why are these changes needed?

To improve the granularity of tests and independently test Ray Core from
the Ray libraries, we want the libraries to use stable APIs from core.
To that end, we want to remove all usages of `ray._private` from the
libraries and move any common utilities that are required but aren't
true public APIs for ray into the `ray._common` directory.

This PR moves all the `_private.utils` being used in the libraries and
need not be in `_private` to the `_common`. And thus replacing all
usages of `ray._private.utils` in the libraries by `ray._common.utils`.

NOTE: No new function has been added. This PR is just a refactor. Unit
tests have been added in `_common/tests/test_utils.py` for some of the
utilities that did not have any tests.

---------

Signed-off-by: Sagar Sumit <sagarsumit09@gmail.com>