[Tune][Air] Fix MLflowLoggerCallback to enable its use with PBT (#27783)
Why are these changes needed?

In the current MLflowLoggerCallback implementation, the function to store parameters in MLflow (mlflow_util.log_param) is only used in log_trial_start.

The problem is that PBT explores various parameters. The initially stored parameters are almost always different from the optimal parameters.

We need to be able to verify not only the initial parameters but also the optimal parameters in MLflow. The current implementation only allows us to verify the initial parameters.

This contribution modifies the MLflowLoggerCallback function to enable the storage of both the final and initial parameters in MLflow.

Related issue number
"Open #27783"

Checks
  - [x] I've signed off every commit(by using the -s flag, i.e., git commit -s) in this PR.
  - [x]  I've run scripts/format.sh to lint the changes in this PR.
 - [  ] I've included any doc changes needed for https://docs.ray.io/en/master/.
 - [  ] I've added any new APIs to the API Reference. For example, if I added a method in Tune, I've added it in doc/source/tune/api/ under the corresponding .rst file.
 I've made sure the tests are passing. Note that there might be a few flaky tests, see the recent failures at https://flakey-tests.ray.io/

Testing Strategy
- [x] Unit tests
- [  ] Release tests
- [  ] This PR is not tested :(

Signed-off-by: Sung Joon Park <sjp611@gmail.com>