[core][telemetry/07] support counter metric on worker side (#53418)
This is a series of PR to migrate metric collection from opencencus to
openlemetry. For context on the existing components, see
https://github.com/ray-project/ray/pull/53098.

---------------

This PR
- Support Counter metric for `open_telemetry_metric_recorder.cc`
- Add a unittest to make sure it works
- This feature is not used anywhere yet, since the counter part
`open_telemetry_metric_recorder.py` needs to be able to consume the
counter metric as well, if this part is used. I'll add that in the next
PR

Test:
- CI

---------

Signed-off-by: can <can@anyscale.com>
Signed-off-by: Cuong Nguyen <128072568+can-anyscale@users.noreply.github.com>
Co-authored-by: Dhyey Shah <dhyey2019@gmail.com>